migrate lib to src/
