--- example/lib/components/WebView/index.android.js
@@ -1,1 +1,0 @@
-export {WebView as default} from "react-native-webview";
