--- example/lib/components/WebView/index.ios.js
@@ -1,1 +1,0 @@
-export {WebView as default} from "react-native-webview";
