--- example/lib/components/WebView/index.web.js
@@ -1,1 +1,0 @@
-export {WebView as default} from "react-native-web-webview";
