--- example/lib/components/index.js
@@ -1,2 +1,0 @@
-export { default as WebView } from "./WebView";
-export { default as TorusModal } from "./TorusModal";
