--- example/lib/contexts/index.js
@@ -1,1 +1,0 @@
-export {default as TorusContext} from "./TorusContext";
