--- example/lib/hooks/index.js
@@ -1,2 +1,0 @@
-export {default as useTorus} from "./useTorus";
-export {default as useKeyPair} from "./useKeyPair";
