--- example/lib/hooks/useSecureStorage/index.android.js
@@ -1,1 +1,0 @@
-export {default} from "./useSecureStorageNative";
