--- example/lib/hooks/useSecureStorage/index.ios.js
@@ -1,1 +1,0 @@
-export {default} from "./useSecureStorageNative";
