--- example/lib/hooks/useSecureStorage/index.web.js
@@ -1,1 +1,0 @@
-export {default} from "./useSecureStorageWeb";
