--- example/lib/hooks/useTorus.js
@@ -1,5 +1,0 @@
-import {useContext} from "react";
-
-import {TorusContext} from "../contexts";
-
-export default () => useContext(TorusContext);
