--- example/lib/index.js
@@ -1,3 +1,0 @@
-export {WebView} from "./components";
-export {useTorus, useKeyPair} from "./hooks";
-export {TorusProvider as default} from "./providers";
