--- example/lib/providers/index.js
@@ -1,1 +1,0 @@
-export {default as TorusProvider} from "./TorusProvider";
