+++ src/components/TorusModal.js
@@ -1,4 +1,4 @@
-import React, { useCallback } from "react";
+import React, { useCallback, useState, useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import ModalBox from "react-native-modalbox";
 import { StyleSheet } from "react-native";
@@ -7,9 +7,14 @@ import { typeCheck } from "type-check";
 import { WebView } from ".";
 
 function TorusModal({ visible, source, onDismiss, onAuthResult }) {
+  const ref = useRef();
+  const [verify, setVerify] = useState(null);
+  const [trigger, setTrigger] = useState(null);
+  const [data, setData] = useState(null);
+  const [key, setKey] = useState(Math.random);
+
   const onMessage = useCallback(
     ({ nativeEvent: { data } }) => {
-      console.warn(data);
       try {
         const e = JSON.parse(data);
         if (typeCheck("{type:String,...}", e)) {
@@ -17,19 +22,62 @@ function TorusModal({ visible, source, onDismiss, onAuthResult }) {
           if (type === "torus-auth") {
             const { data } = extras;
             return onAuthResult(data);
+          } else if (type === "torus-verifier-result") {
+            const { result } = extras;
+            return setVerify(result);
+          } else if (type === "torus-trigger-auth") {
+            const { trigger } = extras;
+            return setTrigger(trigger);
           }
           throw new Error(`Unexpected type, ${type}.`);
         }
       } catch (e) {}
     },
-    [onAuthResult],
+    [onAuthResult, setVerify, ref, setTrigger],
   );
+
+  useEffect(
+    () => {
+      if (!!trigger && !!verify) {
+        const { ...data } = { ...trigger, ...verify };
+        setVerify(null);
+        setTrigger(null);
+        setData(data);
+        setKey(Math.random);
+      }
+    },
+    [trigger, verify, ref, setVerify, setTrigger, setData, setKey],
+  );
+
+  const onLoadEnd = useCallback(
+    () => {
+      if (!!data) {
+        setData(null);
+        ref.current.injectJavaScript(
+          `
+setTimeout(
+  () => window.__REACT_TORUS_TRIGGER_VERIFY__(${JSON.stringify(data)}),
+  1000, // TODO: This needs to be dynamic.
+);
+true;
+          `.trim(),
+        );
+      }
+    },
+    [data, setData, ref],
+  );
+  const { uri } = source;
   return (
     <ModalBox style={StyleSheet.absoluteFill} isOpen={visible} onClosed={onDismiss}>
+      {/* TODO: Find a way to preload the View. */}
       <WebView
+        key={key}
+        ref={ref}
         style={StyleSheet.absoluteFill}
+        onLoadEnd={onLoadEnd}
         source={source}
         onMessage={onMessage}
+        userAgent="Mozilla/5.0 (X11; Linux x86_64; rv:10.0) Gecko/20100101 Firefox/10.0"
       />
     </ModalBox>
   );
